Restructured unit tests Removed ss58, moved to substrate-interface
